manage liquidity from calculator
