+++ packages/client/src/styles/burger-menu.scss
@@ -125,7 +125,7 @@ Note: Beware of modifying this element as it can break the animations - you shou
     font-family: 'Barlow Condensed';
     text-align: center;
     text-transform: uppercase;
-    margin-bottom: 3 rem;
+    margin-bottom: 2rem;
 }
 
 .side-menu-connected {
